+++ AndroidEmotionCalculator/AndroidEmotionCalculator.csproj
@@ -60,15 +60,15 @@
     <Reference Include="System.Numerics.Vectors" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Fragments\CalendarFragment.cs" />
-    <Compile Include="Fragments\GamesFragment.cs" />
-    <Compile Include="Fragments\MusicFragment.cs" />
+    <Compile Include="Fragments\Items\CalendarFragment.cs" />
+    <Compile Include="Fragments\Items\GamesFragment.cs" />
+    <Compile Include="Fragments\Items\MusicFragment.cs" />
+    <Compile Include="Fragments\MainFragmentManager.cs" />
     <Compile Include="Fragments\ScreenAdapter.cs" />
-    <Compile Include="Fragments\ShopFragment.cs" />
+    <Compile Include="Fragments\Items\ShopFragment.cs" />
     <Compile Include="MainActivity.cs" />
     <Compile Include="Resources\Resource.designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ShopActivity.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Resources\AboutResources.txt" />
@@ -138,6 +138,7 @@
       <SubType>Designer</SubType>
     </AndroidResource>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\Android\Xamarin.Android.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
     Other similar extension points exist, see Microsoft.Common.targets.
