+++ AndroidEmotionCalculator/Fragments/Items/CalendarFragment.cs
@@ -2,13 +2,13 @@
 using Android.Support.V4.App;
 using Android.Views;
 
-namespace AndroidEmotionCalculator.Fragments
+namespace AndroidEmotionCalculator.Fragments.Items
 {
-    class MusicFragment : Fragment
+    class CalendarFragment : Fragment
     {
         public override View OnCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
         {
-            return inflater.Inflate(Resource.Layout.music, container, false);
+            return inflater.Inflate(Resource.Layout.calendar, container, false);
         }
     }
 }
