+++ AndroidEmotionCalculator/Fragments/Items/MusicFragment.cs
@@ -1,15 +1,14 @@
-﻿
-using Android.OS;
+﻿using Android.OS;
 using Android.Support.V4.App;
 using Android.Views;
 
-namespace AndroidEmotionCalculator.Fragments
+namespace AndroidEmotionCalculator.Fragments.Items
 {
-    class ShopFragment : Fragment
+    class MusicFragment : Fragment
     {
         public override View OnCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
         {
-            return inflater.Inflate(Resource.Layout.shop, container, false);
+            return inflater.Inflate(Resource.Layout.music, container, false);
         }
     }
 }
