+++ AndroidEmotionCalculator/Fragments/Items/ShopFragment.cs
@@ -1,14 +1,15 @@
-﻿using Android.OS;
+﻿
+using Android.OS;
 using Android.Support.V4.App;
 using Android.Views;
 
-namespace AndroidEmotionCalculator.Fragments
+namespace AndroidEmotionCalculator.Fragments.Items
 {
-    class CalendarFragment : Fragment
+    class ShopFragment : Fragment
     {
         public override View OnCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
         {
-            return inflater.Inflate(Resource.Layout.calendar, container, false);
+            return inflater.Inflate(Resource.Layout.shop, container, false);
         }
     }
 }
